backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
pydantic==2.7.1
paho-mqtt==1.6.1
apscheduler==3.10.4
httpx==0.27.0
python-dotenv==1.0.1
